+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "site",
-	"version": "0.0.1",
+	"version": "0.2.0",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
