+++ src/routes/+layout.svelte
@@ -3,6 +3,7 @@
 	import { page } from '$app/stores';
 	import Footer from '$lib/components/Footer.svelte';
 	import Header from '$lib/components/Header.svelte';
+
 	let title: string;
 
 	afterNavigate(({ to }) => {
