+++ src/routes/+page.svelte
@@ -1,33 +1,8 @@
 <script lang="ts">
 	import GitHubAvatar from '../lib/components/GitHubAvatar.svelte';
+	import type { PageData } from './$types';
 
-	const founders: { displayName: string; twitter: string; githubAuthor: string }[] = [
-		{
-			displayName: 'Domenik Reitzner',
-			twitter: 'https://twitter.com/reitznerdomenik',
-			githubAuthor: 'dreitzner'
-		},
-		{
-			displayName: 'Ermin Čeliković',
-			twitter: 'https://mobile.twitter.com/celikovic',
-			githubAuthor: 'ermincelikovic'
-		},
-		{
-			displayName: 'Ivan Hofer',
-			twitter: 'https://twitter.com/ivanhofer_',
-			githubAuthor: 'ivanhofer'
-		},
-		{
-			displayName: 'Jan-Dennis Drenkhahn',
-			twitter: 'https://twitter.com/crashoverried',
-			githubAuthor: 'stillday'
-		},
-		{
-			displayName: 'Jean-Yves Couët',
-			twitter: 'https://twitter.com/jycouet',
-			githubAuthor: 'jycouet'
-		}
-	];
+	export let data: PageData;
 </script>
 
 <h1>Welcome to Svelte Society Austria</h1>
@@ -36,7 +11,7 @@
 <h2>Founding members</h2>
 
 <ul>
-	{#each founders as { displayName, twitter, githubAuthor }}
+	{#each data.founders as { displayName, twitter, githubAuthor }}
 		<li>
 			<div class="founder">
 				<GitHubAvatar {githubAuthor} size={70} />
